Function rendering::material::cos_and_n

pub fn cos_and_n(
    ray: &Ray,
    normal: Vector3D,
    refraction_index: Float
) -> (Float, Float, Float, Option<Float>)

Calculates the parameters necessary for calculating the Fresnel’s equations. cos2—i.e., the cosine of the angle between the normal and the transmitted ray—is wrapped in an Option because it does not exist if the angle of incidence is larger than the critical angle.

Example

use geometry3d::{Point3D,Vector3D, Ray3D};
use rendering::{Spectrum, Ray};
use rendering::material::cos_and_n;
use rendering::interaction::Interaction;

let mat_refraction_index = 1.52;
let normal = Vector3D::new(0., 0., 1.);
let ray = Ray{
    geometry: Ray3D{
        origin: Point3D::new(0., 0., 1.),
        direction: Vector3D::new(0., 1., -2.).get_normalized()
    },
    .. Ray::default()
};
let (n1, cos1, n2, cos2) = cos_and_n(&ray, normal, mat_refraction_index);