Module rendering::material::specular

Functions

cos_and_n

Calculates the parameters necessary for calculating the Fresnel’s equations. cos2—i.e., the cosine of the angle between the normal and the transmitted ray—is wrapped in an Option because it does not exist if the angle of incidence is larger than the critical angle.

eval_mirror_bsdf

Evaluates the mirror BSDf

fresnel_reflectance

fresnel_te

Fresnel Coefficient for TE-Polarized Light (i.e., perpendicular), according to PBR-book

fresnel_tm

Fresnel Coefficient for TM-Polarized Light (i.e., parallel), according to PBR-book

fresnel_transmission_dir

Calculates the direction of the transmision

mirror_bsdf

Calculates the Mirror BSDF and modifies the given ray so that it now points in that direction

mirror_direction

Calculates the purely specular reflection direction.