Trait rendering::material::Material

pub trait Material {
    fn id(&self) -> &str;

    fn colour(&self) -> Spectrum;

    fn sample_bsdf(
        &self, 
        normal: Vector3D, 
        e1: Vector3D, 
        e2: Vector3D, 
        intersection_pt: Point3D, 
        ray: &mut Ray, 
        rng: &mut RandGen
    ) -> (f64, bool);

    fn eval_bsdf(
        &self, 
        normal: Vector3D, 
        e1: Vector3D, 
        e2: Vector3D, 
        ray: &Ray, 
        vout: Vector3D
    ) -> f64;

    fn emits_direct_light(&self) -> bool { ... }

    fn emits_light(&self) -> bool { ... }

    fn specular_only(&self) -> bool { ... }

    fn get_possible_paths(
        &self, 
        _normal: &Vector3D, 
        _intersection_pt: &Point3D, 
        _ray: &Ray
    ) -> [Option<(Ray, f64, f64)>; 2] { ... }
}

Required methods

fn id(&self) -> &str

Returns an id, for debugging and error reporting purposes

fn colour(&self) -> Spectrum

Retrieves the Colour of the material. This will usually represent the values that will multiply the different elements of the Spectrum. E.g., the reflectance values.

fn sample_bsdf(
&self,
normal: Vector3D,
e1: Vector3D,
e2: Vector3D,
intersection_pt: Point3D,
ray: &mut Ray,
rng: &mut RandGen
) -> (f64, bool)

Samples the bsdf (returned by modifying the given Ray). Returns the value of the BSDF in that direction, and a boolean indicating whether this is a specular or a diffuse interaction

fn eval_bsdf(
&self,
normal: Vector3D,
e1: Vector3D,
e2: Vector3D,
ray: &Ray,
vout: Vector3D
) -> f64

Evaluates a BSDF based on an input and outpt directions

Provided methods

fn emits_direct_light(&self) -> bool

Should this material be tested for direct illumination?

fn emits_light(&self) -> bool

Should this material emits light

fn specular_only(&self) -> bool

Does this material scatter (e.g., like Plastic) or does it only transmit/reflects specularly (e.g., like Mirror)?

Defaults to false

fn get_possible_paths(
&self,
_normal: &Vector3D,
_intersection_pt: &Point3D,
_ray: &Ray
) -> [Option<(Ray, f64, f64)>; 2]