Struct rendering::rand::SmallRng

pub struct SmallRng(Xoshiro256PlusPlus);

A small-state, fast non-crypto PRNG

SmallRng may be a good choice when a PRNG with small state, cheap initialization, good statistical quality and good performance are required. Note that depending on the application, StdRng may be faster on many modern platforms while providing higher-quality randomness. Furthermore, SmallRng is not a good choice when:

Security against prediction is important. Use StdRng instead.

Seeds with many zeros are provided. In such cases, it takes SmallRng about 10 samples to produce 0 and 1 bits with equal probability. Either provide seeds with an approximately equal number of 0 and 1 (for example by using SeedableRng::from_entropy or SeedableRng::seed_from_u64), or use StdRng instead.

The algorithm is deterministic but should not be considered reproducible due to dependence on platform and possible replacement in future library versions. For a reproducible generator, use a named PRNG from an external crate, e.g. rand_xoshiro or rand_chacha. Refer also to The Book.

The PRNG algorithm in SmallRng is chosen to be efficient on the current platform, without consideration for cryptography or security. The size of its state is much smaller than StdRng. The current algorithm is Xoshiro256PlusPlus on 64-bit platforms and Xoshiro128PlusPlus on 32-bit platforms. Both are also implemented by the rand_xoshiro crate.

Examples

Initializing SmallRng with a random seed can be done using SeedableRng::from_entropy:

use rand::{Rng, SeedableRng};
use rand::rngs::SmallRng;

// Create small, cheap to initialize and fast RNG with a random seed.
// The randomness is supplied by the operating system.
let mut small_rng = SmallRng::from_entropy();

When initializing a lot of SmallRng’s, using thread_rng can be more efficient:

use rand::{SeedableRng, thread_rng};
use rand::rngs::SmallRng;

// Create a big, expensive to initialize and slower, but unpredictable RNG.
// This is cached and done only once per thread.
let mut thread_rng = thread_rng();
// Create small, cheap to initialize and fast RNGs with random seeds.
// One can generally assume this won't fail.
let rngs: Vec<SmallRng> = (0..10)
    .map(|_| SmallRng::from_rng(&mut thread_rng).unwrap())
    .collect();

Tuple Fields
0: Xoshiro256PlusPlus