Struct rendering::scene::Scene

pub struct Scene {
    pub triangles: Vec<Triangle>,
    pub normals: Vec<(Vector3D, Vector3D, Vector3D)>,
    pub front_material_indexes: Vec<usize>,
    pub back_material_indexes: Vec<usize>,
    pub materials: Vec<Material>,
    pub lights: Vec<Object>,
    pub distant_lights: Vec<Object>,
    pub accelerator: Option<BoundingVolumeTree>,
    pub sky_colour: Option<Spectrum<{ crate::N_CHANNELS }>>,
    pub sky: Option<Box<dyn Fn(Vector3D) -> Float + Sync>>,
}

Fields
triangles: Vec<Triangle>

The Triangles in the scene that are not tested directly for shadow (e.g., non-luminous objects and diffuse light)

normals: Vec<(Vector3D, Vector3D, Vector3D)>

The normal of each vertex of each triangle.

front_material_indexes: Vec<usize>back_material_indexes: Vec<usize>materials: Vec<Material>

The materials in the scene

lights: Vec<Object>

A vector of Light objects that are considered sources of direct light. The objects here are also in the objects part.

distant_lights: Vec<Object>

A vector of distant Light objects that are considered sources of direct light

accelerator: Option<BoundingVolumeTree>

The acceleration structure that helps trace rays.

This needs to be build through the build_accelerator function.

sky_colour: Option<Spectrum<{ crate::N_CHANNELS }>>

The colour of the sky, normalized

sky: Option<Box<dyn Fn(Vector3D) -> Float + Sync>>

A function returning the diffuse Sky brightness (i.e., without the sun) The sun should be added separately. Alternatively, you can use the add_perez_sky function

Implementations

impl Scene

pub fn from_simple_model(model: &SimpleModel, wavelength: Wavelengths) -> Self

Creates a new Scene from a SimpleModel. The enum Wavelengths can be used to create a Visible or a Solar model, for calculating Lighting or Solar Radiation, respectively.

pub fn new() -> Self

Creates an empty scene

pub fn add_perez_sky(
&mut self,
date: Date,
latitude: Float,
longitude: Float,
standard_meridian: Float,
diffuse_horizontal_irrad: Float,
direct_normal_irrad: Float
)

Adds the elements describing a Perez sky to the scene. The angles of Latitude, Longitude and Standard meridian should come in Degrees

pub fn build_accelerator(&mut self)

pub fn rebuild_accelerator(&mut self)

Builds the accelerator

pub fn count_all_lights(&self) -> usize

Returns the number of total lights; that is, those in the lighs field and those in the distant_lights one

pub fn cast_ray(&self, ray: &mut Ray, node_aux: &mut Vec<usize>) -> Option<usize>

Casts a [Ray3D] and returns an Option<usize> indicating the index of the first primitive hit by the ray, if any. The ray passed will now contain the Interaction

pub fn unobstructed_distance(
&self,
ray: &Ray3D,
distance_squared: Float,
node_aux: &mut Vec<usize>
) -> bool

Checks whether a [Ray3D] can travel a certain distance without hitting any surface

pub fn push_material(&mut self, material: Material) -> usize

Pushes a Material to the Scene and return its position in the materials Vector.

pub fn push_object(
&mut self,
front_material_index: usize,
back_material_index: usize,
primitive: Primitive
)

Pushes a Primitive object into the Scene

If the Primitive is made of a light-emmiting Material, then it will be added twice: One to the normal scene, and then another to the list of light sources.

impl Scene

pub fn add_from_obj(
&mut self,
filename: String,
front_material_index: usize,
back_material_index: usize
)

Reads a Radiance file and builds a scene.

pub fn add_from_obj_source(
&mut self,
source: &[u8],
front_material_index: usize,
back_material_index: usize
)

Creates a scene from a slice of bytes read from a Radiance file

impl Scene

pub fn from_radiance(filename: String) -> Self

Reads a Radiance file and builds a scene.

pub fn from_radiance_source(source: &[u8]) -> Self

Creates a scene from a slice of bytes read from a Radiance file