Module rendering::triangle

Structs

Intersection

Functions

baricentric_coorinates 🔒

Tests the intersection between a Ray3D and a Triangle. Returns the the point of intersection, and the u and v baricentric coordinates of the intersection point.

cross 🔒

dot 🔒

mesh_sphere

mesh_triangle

Transforms a Triangle3D and transforms it into a Vec<Triangle> and their respective normals

new_info

simple_triangle_intersect

Intersects a Ray3D and a Triangle, returning the Point3D of intersection

simple_triangle_intersect_pack

Intersects a Ray3D and a pack (i.e., &[]) of Triangle, returning the index of the intersected Triangle within the pack, and the Point3D of intersection

triangle_area

triangle_intersect

Intersects a Ray3D and a Triangle, returning the [IntersectionInfo] (or None if they don’t intersect)

triangle_intersect_pack

Intersects a Ray3D and a pack (i.e., &[]) of Triangle, returning the index of the intersected Triangle within the pack, and its [IntersectionInfo] (or None if they don’t intersect)

triangle_pack_baricentric_coorinates

Tests the intersection between a Ray3D and a pack (i.e., &[]) of Triangle. Returns the index of the intersected triangle within the pack, the point of intersection, and the u and v baricentric coordinates of the intersection point.

triangle_solid_angle_pdf

world_bounds

Gets the BBox of a Triangle

Type Definitions

Triangle

The smallest definition of a Triangle I could think of