Trait rendering::material::Material

pub trait Material {
    fn colour(&self) -> Spectrum;

    fn bsdf(&self, vin: Vector3D, normal: Vector3D, vout: Vector3D) -> f64;

    fn sample_bsdf(
        &self, 
        rng: &mut ThreadRng, 
        vout: Vector3D, 
        normal: Vector3D
    ) -> Vector3D;

    fn emits_direct_light(&self) -> bool { ... }

    fn emits_light(&self) -> bool { ... }

    fn specular_only(&self) -> bool { ... }
}

Required methods

fn colour(&self) -> Spectrum

Retrieves the Colour of the material. This will usually represent the values that will multiply the different elements of the Spectrum. E.g., the reflectance values.

fn bsdf(&self, vin: Vector3D, normal: Vector3D, vout: Vector3D) -> f64

Gets the BSDF’s value for a certain combination of Vin, Vout and Normal [Vector3D]s.

fn sample_bsdf(
&self,
rng: &mut ThreadRng,
vout: Vector3D,
normal: Vector3D
) -> Vector3D

Gets a sample associated to the bsdf

Provided methods

fn emits_direct_light(&self) -> bool

Should this material be tested for direct illumination?

Defaults to false

fn emits_light(&self) -> bool

Should this material emits light

Defaults to false

fn specular_only(&self) -> bool

Does this material scatter (e.g., like Plastic) or does it only transmit/reflects specularly (e.g., like Mirror)?

Defaults to false