Trait rendering::material::Material

pub trait Material: Sync + Send {
    fn colour(&self) -> Spectrum;

    fn bsdf_sampler(
        &self, 
        shading_info: ShadingInfo
    ) -> Box<dyn Fn(Vector3D, &mut RandGen) -> (Vector3D, f64)>;

    fn bsdf_evaluator(
        &self, 
        shading_info: ShadingInfo
    ) -> Box<dyn Fn(Vector3D, Vector3D) -> f64>;

    fn emits_direct_light(&self) -> bool { ... }

    fn emits_light(&self) -> bool { ... }

    fn specular_only(&self) -> bool { ... }
}

Required methods

fn colour(&self) -> Spectrum

Retrieves the Colour of the material. This will usually represent the values that will multiply the different elements of the Spectrum. E.g., the reflectance values.

fn bsdf_sampler(
&self,
shading_info: ShadingInfo
) -> Box<dyn Fn(Vector3D, &mut RandGen) -> (Vector3D, f64)>

Gets an iterator that samples the BSDF of the material according to a certain ShadingInfo. The closure receives a Vector3D corresponding to the incident direction in world’s coordinates, and returns the outgoing direction as well as the probability of going in that direction.

Note

When implementing forward/backard ray-tracing, modifications will have to be made to this function.

fn bsdf_evaluator(
&self,
shading_info: ShadingInfo
) -> Box<dyn Fn(Vector3D, Vector3D) -> f64>

Gets a closure that allows calculating and sampling the BSDF of the material according to a certain ShadingInfo. The closure receives two Vector3Ds, corresponding to the incident and outgoing directions in world’s coordinates, and returns the value of the BSDF accordingly.

Note

When implementing forward/backard ray-tracing, modifications will have to be made to this function.

Provided methods

fn emits_direct_light(&self) -> bool

Should this material be tested for direct illumination?

Defaults to false

fn emits_light(&self) -> bool

Should this material emits light

Defaults to false

fn specular_only(&self) -> bool

Does this material scatter (e.g., like Plastic) or does it only transmit/reflects specularly (e.g., like Mirror)?

Defaults to false